refactor(encoding): Phase 1e Stage D batch 20a — host BT walker on MatchTable
Mirror `search_depth`, `is_btultra2`, and `uses_bt` from `HcMatcher` /
`HcParseMode` onto new `MatchTable` fields during `configure()`. The
former `HcMatchGenerator::is_btultra2()` and `uses_bt_matchfinder()`
accessors now consult those flags so the BT walker can read its
dispatch state without going back through the outer generator.

Move the per-position BT walker (`bt_insert_step_no_rebase` plus four
SIMD variants, ten methods total when counted with
`bt_insert_and_collect_matches`) from `impl BtMatcher` onto
`impl MatchTable`. The body macros only ever touched `$table` /
`$search_depth`, never the matcher itself — the move just drops the
no-op `let _ = self;` lines and changes the receiver from
`&self: BtMatcher + table: &mut MatchTable` to `&mut self: MatchTable`,
reading `self.search_depth` directly.

Replace BT-side `replay_history_for_rebase` (previously on BtMatcher)
with `MatchTable::replay_history_for_rebase_bt`. `rebase_positions_cold`
now dispatches via `self.table.uses_bt` and calls the appropriate
MatchTable replay variant directly — the split borrow between
`&self.bt` and `&mut self.table` is gone.

Every prior call site (`bt_update_tree_until_*` SIMD variants,
`bt_insert_sparse_incompressible_block`, the
`collect_optimal_candidates_initialized_body!` macro) is rewritten:
`self.bt.bt_insert_step_no_rebase*(&mut self.table, self.hc.search_depth, ...)`
becomes `self.table.bt_insert_step_no_rebase*(...)`, and
`$self.bt.$bt_insert(&mut $self.table, $self.hc.search_depth, ...)`
becomes `$self.table.$bt_insert(...)`.

Two test sites (`btultra_and_btultra2_both_keep_dictionary_candidates`)
that bypass `configure()` are taught to write through the new flags
when they mutate `parse_mode` directly.

Verification:
  cargo build -p structured-zstd
  cargo build -p structured-zstd --no-default-features
  cargo clippy --all-targets -p structured-zstd -- -D warnings
  cargo nextest run -p structured-zstd --lib    # 381/381, level22 ratio gate green
