+++ zstd/src/encoding/bt/mod.rs
@@ -232,176 +232,6 @@ impl BtMatcher {
         )
     }
 
-    /// Cross-platform dispatcher for the per-position BT walker. The
-    /// per-iteration `count_match_from_indices` symbol inlines under
-    /// the kernel-specific `target_feature` umbrella so the entire
-    /// walk runs as one straight-line hot path.
-    #[inline(always)]
-    pub(crate) fn bt_insert_step_no_rebase(
-        &self,
-        table: &mut MatchTable,
-        search_depth: usize,
-        abs_pos: usize,
-        current_abs_end: usize,
-        target_abs: usize,
-    ) -> usize {
-        // SAFETY: each branch verifies the target_feature requirement
-        // of the callee — aarch64 NEON is baseline; x86 AVX2/BMI2 and
-        // SSE4.2 are selected only when the runtime detector reports
-        // them present.
-        #[cfg(all(target_arch = "aarch64", target_endian = "little"))]
-        unsafe {
-            self.bt_insert_step_no_rebase_neon(
-                table,
-                search_depth,
-                abs_pos,
-                current_abs_end,
-                target_abs,
-            )
-        }
-        #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-        {
-            use crate::encoding::fastpath::{FastpathKernel, select_kernel};
-            match select_kernel() {
-                FastpathKernel::Avx2Bmi2 => unsafe {
-                    self.bt_insert_step_no_rebase_avx2_bmi2(
-                        table,
-                        search_depth,
-                        abs_pos,
-                        current_abs_end,
-                        target_abs,
-                    )
-                },
-                FastpathKernel::Sse42 => unsafe {
-                    self.bt_insert_step_no_rebase_sse42(
-                        table,
-                        search_depth,
-                        abs_pos,
-                        current_abs_end,
-                        target_abs,
-                    )
-                },
-                FastpathKernel::Scalar => self.bt_insert_step_no_rebase_scalar(
-                    table,
-                    search_depth,
-                    abs_pos,
-                    current_abs_end,
-                    target_abs,
-                ),
-            }
-        }
-        #[cfg(not(any(
-            all(target_arch = "aarch64", target_endian = "little"),
-            target_arch = "x86",
-            target_arch = "x86_64"
-        )))]
-        {
-            self.bt_insert_step_no_rebase_scalar(
-                table,
-                search_depth,
-                abs_pos,
-                current_abs_end,
-                target_abs,
-            )
-        }
-    }
-
-    /// NEON umbrella variant of the BT walker. Body inlines via macro.
-    ///
-    /// # Safety
-    /// AArch64 with NEON (baseline).
-    #[cfg(all(target_arch = "aarch64", target_endian = "little"))]
-    #[target_feature(enable = "neon")]
-    pub(crate) unsafe fn bt_insert_step_no_rebase_neon(
-        &self,
-        table: &mut MatchTable,
-        search_depth: usize,
-        abs_pos: usize,
-        current_abs_end: usize,
-        target_abs: usize,
-    ) -> usize {
-        let _ = self;
-        crate::bt_insert_step_no_rebase_body!(
-            table,
-            search_depth,
-            abs_pos,
-            current_abs_end,
-            target_abs,
-            crate::encoding::fastpath::neon::count_match_from_indices
-        )
-    }
-
-    /// SSE4.2 umbrella BT walker.
-    ///
-    /// # Safety
-    /// x86/x86_64 with SSE4.2.
-    #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-    #[target_feature(enable = "sse4.2")]
-    pub(crate) unsafe fn bt_insert_step_no_rebase_sse42(
-        &self,
-        table: &mut MatchTable,
-        search_depth: usize,
-        abs_pos: usize,
-        current_abs_end: usize,
-        target_abs: usize,
-    ) -> usize {
-        let _ = self;
-        crate::bt_insert_step_no_rebase_body!(
-            table,
-            search_depth,
-            abs_pos,
-            current_abs_end,
-            target_abs,
-            crate::encoding::fastpath::sse42::count_match_from_indices
-        )
-    }
-
-    /// AVX2+BMI2 umbrella BT walker.
-    ///
-    /// # Safety
-    /// x86/x86_64 with AVX2 + BMI2.
-    #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-    #[target_feature(enable = "avx2,bmi2")]
-    pub(crate) unsafe fn bt_insert_step_no_rebase_avx2_bmi2(
-        &self,
-        table: &mut MatchTable,
-        search_depth: usize,
-        abs_pos: usize,
-        current_abs_end: usize,
-        target_abs: usize,
-    ) -> usize {
-        let _ = self;
-        crate::bt_insert_step_no_rebase_body!(
-            table,
-            search_depth,
-            abs_pos,
-            current_abs_end,
-            target_abs,
-            crate::encoding::fastpath::avx2_bmi2::count_match_from_indices
-        )
-    }
-
-    /// Scalar fallback BT walker (used on non-AArch64 targets).
-    #[cfg(not(all(target_arch = "aarch64", target_endian = "little")))]
-    pub(crate) fn bt_insert_step_no_rebase_scalar(
-        &self,
-        table: &mut MatchTable,
-        search_depth: usize,
-        abs_pos: usize,
-        current_abs_end: usize,
-        target_abs: usize,
-    ) -> usize {
-        let _ = self;
-        crate::bt_insert_step_no_rebase_body!(
-            table,
-            search_depth,
-            abs_pos,
-            current_abs_end,
-            target_abs,
-            crate::encoding::fastpath::scalar::count_match_from_indices
-        )
-    }
-
     pub(crate) fn new() -> Self {
         Self {
             opt_state: HcOptState::new(),
@@ -442,229 +272,6 @@ impl BtMatcher {
         self.opt_ml_price_generation.clear();
         self.opt_ml_price_stamp = 0;
         self.ldm_sequences.clear();
-    }
-
-    /// Cross-platform entry. Picks the kernel-specific variant so the BT walk
-    /// body executes inside one `target_feature` umbrella and inlines the
-    /// vectorized `count_match_from_indices` directly. See
-    /// `bt_insert_step_no_rebase` for the same dispatcher pattern.
-    ///
-    /// The on-encode hot path bypasses this dispatcher: when invoked from
-    /// `collect_optimal_candidates_initialized_<kernel>` the per-kernel
-    /// variant is called directly so the BT match collection inlines under
-    /// the surrounding umbrella. This entry is kept for external / future
-    /// callers that aren't yet under an umbrella.
-    #[allow(dead_code)]
-    #[allow(clippy::too_many_arguments)]
-    #[inline(always)]
-    pub(crate) fn bt_insert_and_collect_matches(
-        &self,
-        table: &mut MatchTable,
-        search_depth: usize,
-        abs_pos: usize,
-        current_abs_end: usize,
-        profile: HcOptimalCostProfile,
-        min_match_len: usize,
-        best_len_for_skip: &mut usize,
-        out: &mut Vec<MatchCandidate>,
-    ) {
-        // SAFETY: each branch verifies the target_feature requirement of the
-        // callee (see `bt_insert_step_no_rebase` dispatcher).
-        #[cfg(all(target_arch = "aarch64", target_endian = "little"))]
-        unsafe {
-            self.bt_insert_and_collect_matches_neon(
-                table,
-                search_depth,
-                abs_pos,
-                current_abs_end,
-                profile,
-                min_match_len,
-                best_len_for_skip,
-                out,
-            )
-        }
-        #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-        {
-            use crate::encoding::fastpath::{FastpathKernel, select_kernel};
-            match select_kernel() {
-                FastpathKernel::Avx2Bmi2 => unsafe {
-                    self.bt_insert_and_collect_matches_avx2_bmi2(
-                        table,
-                        search_depth,
-                        abs_pos,
-                        current_abs_end,
-                        profile,
-                        min_match_len,
-                        best_len_for_skip,
-                        out,
-                    )
-                },
-                FastpathKernel::Sse42 => unsafe {
-                    self.bt_insert_and_collect_matches_sse42(
-                        table,
-                        search_depth,
-                        abs_pos,
-                        current_abs_end,
-                        profile,
-                        min_match_len,
-                        best_len_for_skip,
-                        out,
-                    )
-                },
-                FastpathKernel::Scalar => self.bt_insert_and_collect_matches_scalar(
-                    table,
-                    search_depth,
-                    abs_pos,
-                    current_abs_end,
-                    profile,
-                    min_match_len,
-                    best_len_for_skip,
-                    out,
-                ),
-            }
-        }
-        #[cfg(not(any(
-            all(target_arch = "aarch64", target_endian = "little"),
-            target_arch = "x86",
-            target_arch = "x86_64"
-        )))]
-        {
-            self.bt_insert_and_collect_matches_scalar(
-                table,
-                search_depth,
-                abs_pos,
-                current_abs_end,
-                profile,
-                min_match_len,
-                best_len_for_skip,
-                out,
-            )
-        }
-    }
-
-    /// NEON-umbrella variant of `bt_insert_and_collect_matches`. Inlines
-    /// `fastpath::neon::count_match_from_indices` via the shared body macro.
-    ///
-    /// # Safety
-    /// AArch64 with NEON (baseline).
-    #[cfg(all(target_arch = "aarch64", target_endian = "little"))]
-    #[target_feature(enable = "neon")]
-    #[allow(clippy::too_many_arguments)]
-    pub(crate) unsafe fn bt_insert_and_collect_matches_neon(
-        &self,
-        table: &mut MatchTable,
-        search_depth: usize,
-        abs_pos: usize,
-        current_abs_end: usize,
-        profile: HcOptimalCostProfile,
-        min_match_len: usize,
-        best_len_for_skip: &mut usize,
-        out: &mut Vec<MatchCandidate>,
-    ) {
-        let _ = self;
-        crate::bt_insert_and_collect_matches_body!(
-            table,
-            search_depth,
-            abs_pos,
-            current_abs_end,
-            profile,
-            min_match_len,
-            best_len_for_skip,
-            out,
-            crate::encoding::fastpath::neon::count_match_from_indices,
-        )
-    }
-
-    /// SSE4.2 umbrella variant of `bt_insert_and_collect_matches`.
-    ///
-    /// # Safety
-    /// x86/x86_64 with SSE4.2.
-    #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-    #[target_feature(enable = "sse4.2")]
-    #[allow(clippy::too_many_arguments)]
-    pub(crate) unsafe fn bt_insert_and_collect_matches_sse42(
-        &self,
-        table: &mut MatchTable,
-        search_depth: usize,
-        abs_pos: usize,
-        current_abs_end: usize,
-        profile: HcOptimalCostProfile,
-        min_match_len: usize,
-        best_len_for_skip: &mut usize,
-        out: &mut Vec<MatchCandidate>,
-    ) {
-        let _ = self;
-        crate::bt_insert_and_collect_matches_body!(
-            table,
-            search_depth,
-            abs_pos,
-            current_abs_end,
-            profile,
-            min_match_len,
-            best_len_for_skip,
-            out,
-            crate::encoding::fastpath::sse42::count_match_from_indices,
-        )
-    }
-
-    /// AVX2+BMI2 umbrella variant of `bt_insert_and_collect_matches`.
-    ///
-    /// # Safety
-    /// x86/x86_64 with AVX2 and BMI2.
-    #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-    #[target_feature(enable = "avx2,bmi2")]
-    #[allow(clippy::too_many_arguments)]
-    pub(crate) unsafe fn bt_insert_and_collect_matches_avx2_bmi2(
-        &self,
-        table: &mut MatchTable,
-        search_depth: usize,
-        abs_pos: usize,
-        current_abs_end: usize,
-        profile: HcOptimalCostProfile,
-        min_match_len: usize,
-        best_len_for_skip: &mut usize,
-        out: &mut Vec<MatchCandidate>,
-    ) {
-        let _ = self;
-        crate::bt_insert_and_collect_matches_body!(
-            table,
-            search_depth,
-            abs_pos,
-            current_abs_end,
-            profile,
-            min_match_len,
-            best_len_for_skip,
-            out,
-            crate::encoding::fastpath::avx2_bmi2::count_match_from_indices,
-        )
-    }
-
-    /// Scalar fallback used on non-AArch64 targets.
-    #[cfg(not(all(target_arch = "aarch64", target_endian = "little")))]
-    #[allow(clippy::too_many_arguments)]
-    pub(crate) fn bt_insert_and_collect_matches_scalar(
-        &self,
-        table: &mut MatchTable,
-        search_depth: usize,
-        abs_pos: usize,
-        current_abs_end: usize,
-        profile: HcOptimalCostProfile,
-        min_match_len: usize,
-        best_len_for_skip: &mut usize,
-        out: &mut Vec<MatchCandidate>,
-    ) {
-        let _ = self;
-        crate::bt_insert_and_collect_matches_body!(
-            table,
-            search_depth,
-            abs_pos,
-            current_abs_end,
-            profile,
-            min_match_len,
-            best_len_for_skip,
-            out,
-            crate::encoding::fastpath::scalar::count_match_from_indices,
-        )
     }
 
     /// Donor parity: `ZSTD_optLdm_skipRawSeqStoreBytes`. Fast-forward the
@@ -841,28 +448,6 @@ impl BtMatcher {
     pub(crate) fn prepare_ldm_candidates(&mut self, current_abs_start: usize, current_len: usize) {
         self.ldm_sequences.clear();
         let _ = (current_abs_start, current_len);
-    }
-
-    /// BT-side history replay after [`MatchTable::begin_rebase`].
-    /// Walks `history_start..abs_pos` through the BT walker so the
-    /// pointer-pair table is consistent with the fresh
-    /// `position_base`. `rebuild_end` is fixed at the live
-    /// `history_abs_end()` so the BT inserts don't probe beyond the
-    /// already-mirrored bytes.
-    pub(crate) fn replay_history_for_rebase(
-        &self,
-        table: &mut MatchTable,
-        search_depth: usize,
-        history_start: usize,
-        abs_pos: usize,
-    ) {
-        let rebuild_end = table.history_abs_end();
-        let mut pos = history_start;
-        while pos < abs_pos {
-            let forward =
-                self.bt_insert_step_no_rebase(table, search_depth, pos, rebuild_end, abs_pos);
-            pos = pos.saturating_add(forward.max(1));
-        }
     }
 
     /// Donor parity: `ZSTD_storeSeq` — encode `actual_offset` into the
